backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
aiosqlite
python-dotenv
pydantic
